requirements.txt
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
nornir==3.4.1
pyaml==24.9.0
Pygments==2.17.2
PyYAML==6.0.2
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
setuptools==75.7.0
